+++ be/src/exec/data_sink.cpp
@@ -34,12 +34,11 @@
 #include "runtime/result_file_sink.h"
 #include "runtime/result_sink.h"
 #include "runtime/runtime_state.h"
-
 #include "vec/sink/result_sink.h"
 #include "vec/sink/vdata_stream_sender.h"
+#include "vec/sink/vmysql_table_sink.h"
 #include "vec/sink/vmysql_table_writer.h"
 #include "vec/sink/vtablet_sink.h"
-#include "vec/sink/vmysql_table_sink.h"
 
 namespace doris {
 
@@ -80,7 +79,8 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
 
         // TODO: figure out good buffer size based on size of output row
         if (is_vec) {
-            tmp_sink = new doris::vectorized::VResultSink(row_desc, output_exprs, thrift_sink.result_sink, 4096);
+            tmp_sink = new doris::vectorized::VResultSink(row_desc, output_exprs,
+                                                          thrift_sink.result_sink, 4096);
         } else {
             tmp_sink = new ResultSink(row_desc, output_exprs, thrift_sink.result_sink, 1024);
         }
@@ -111,12 +111,12 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
         break;
     }
     case TDataSinkType::MYSQL_TABLE_SINK: {
-#ifdef DORIS_WITH_MYSQL
         if (!thrift_sink.__isset.mysql_table_sink) {
             return Status::InternalError("Missing data buffer sink.");
         }
         if (is_vec) {
-            doris::vectorized::VMysqlTableSink* vmysql_tbl_sink = new doris::vectorized::VMysqlTableSink(pool, row_desc, output_exprs);
+            doris::vectorized::VMysqlTableSink* vmysql_tbl_sink =
+                    new doris::vectorized::VMysqlTableSink(pool, row_desc, output_exprs);
             sink->reset(vmysql_tbl_sink);
         } else {
             // TODO: figure out good buffer size based on size of output row
@@ -124,10 +124,6 @@ Status DataSink::create_data_sink(ObjectPool* pool, const TDataSink& thrift_sink
             sink->reset(mysql_tbl_sink);
         }
         break;
-#else
-        return Status::InternalError(
-                "Don't support MySQL table, you should rebuild Doris with WITH_MYSQL option ON");
-#endif
     }
     case TDataSinkType::ODBC_TABLE_SINK: {
         if (!thrift_sink.__isset.odbc_table_sink) {