+++ be/src/vec/sink/vmysql_table_writer.cpp
@@ -17,7 +17,7 @@
 
 #include "vec/sink/vmysql_table_writer.h"
 
-#include <mysql/mysql.h>
+#include <mariadb/mysql.h>
 
 #include <sstream>