+++ thirdparty/download-thirdparty.sh
@@ -244,15 +244,6 @@ fi
 cd -
 echo "Finished patching $GTEST_SOURCE"
 
-# mysql patch
-cd $TP_SOURCE_DIR/$MYSQL_SOURCE
-if [ ! -f $PATCHED_MARK ]; then
-    patch -p1 < $TP_PATCH_DIR/mysql-server-mysql-5.7.18.patch
-    touch $PATCHED_MARK
-fi
-cd -
-echo "Finished patching $MYSQL_SOURCE"
-
 # libevent patch
 cd $TP_SOURCE_DIR/$LIBEVENT_SOURCE
 if [ ! -f $PATCHED_MARK ]; then
@@ -305,8 +296,14 @@ echo "Finished patching $LIBRDKAFKA_SOURCE"
 cd $TP_SOURCE_DIR/$AWS_SDK_SOURCE
 if [ ! -f $PATCHED_MARK ]; then
     if [ $AWS_SDK_SOURCE == "aws-sdk-cpp-1.9.211" ]; then
-        wget --no-check-certificate -q https://doris-thirdparty-repo.bj.bcebos.com/thirdparty/aws-crt-cpp-1.9.211.tar.gz
-        tar xzf aws-crt-cpp-1.9.211.tar.gz
+        wget --no-check-certificate -q https://doris-thirdparty-repo.bj.bcebos.com/thirdparty/aws-crt-cpp-1.9.211.tar.gz -O aws-crt-cpp-1.9.211.tar.gz
+        ret="$?"
+        if [ $ret -eq 0 ] ; then
+            tar xzf aws-crt-cpp-1.9.211.tar.gz
+        else
+            bash ./prefetch_crt_dependency.sh
+        fi
+        
     else
         bash ./prefetch_crt_dependency.sh
     fi